Moves ToiOutput out to its own files
